+++ packages/mention/src/MentionSuggestions/MentionSuggestions.tsx
@@ -51,6 +51,7 @@ export interface MentionSuggestionsPubProps {
   >;
   entryComponent?: ComponentType<EntryComponentProps>;
   popoverContainer?: ComponentType<PopperOptions>;
+  renderEmptyPopup?: boolean;
 }
 
 export interface MentionSuggestionsProps extends MentionSuggestionsPubProps {
@@ -393,6 +394,10 @@ export class MentionSuggestions extends Component<MentionSuggestionsProps> {
           />
         ))
       );
+    }
+
+    if (!this.props.renderEmptyPopup && this.props.suggestions.length === 0) {
+      return null;
     }
 
     return (